+++ plugins/editors/intern/editor.cpp
@@ -1,16 +1,16 @@
-#include "editor.hpp"
+#include "../editor.hpp"
 
-#include <menu.hpp>
+#include "menu/menu.hpp"
 
 namespace {
 constinit LOG_HANDLE *logger = nullptr;
 }  // namespace
 
-namespace flow::editors {
+namespace flow::editor {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *handle) {
     logger = handle;
 
     menu::init(host, logger);
 }
-}  // namespace flow::editors
+}  // namespace flow::editor