+++ plugins/editors/intern/menu/intern/menu.cpp
@@ -1,12 +1,12 @@
-#include "menu.hpp"
+#include "../menu.hpp"
 
-#include <object.hpp>
-#include <property.hpp>
+#include "object/object.hpp"
+#include "property/property.hpp"
 
-namespace flow::editors::menu {
+namespace flow::editor::menu {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger) {
     object::init(host, logger);
     property::init(host, logger);
 }
-}  // namespace flow::editors::menu
+}  // namespace flow::editor::menu