+++ plugins/editors/intern/menu/intern/object/intern/object.cpp
@@ -1,10 +1,10 @@
-#include "object.hpp"
+#include "../object.hpp"
 
 #include "text.hpp"
 
-namespace flow::editors::menu::object {
+namespace flow::editor::menu::object {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger) {
-    flow::editors::menu::object::text::init(host, logger);
+    flow::editor::menu::object::text::init(host, logger);
 }
-}  // namespace flow::editors::menu::object
+}  // namespace flow::editor::menu::object