+++ plugins/editors/intern/menu/intern/object/object.hpp
@@ -5,7 +5,7 @@
 #include <logger2.h>
 #include <plugin2.h>
 
-namespace flow::editors::menu::object {
+namespace flow::editor::menu::object {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger);
-}  // namespace flow::editors::menu::object
+}  // namespace flow::editor::menu::object