+++ plugins/modules/intern/hash/hash.hpp
@@ -5,7 +5,7 @@
 #include <logger2.h>
 #include <plugin2.h>
 
-namespace flow::modules::hash {
+namespace flow::module::hash {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger);
-}  // namespace flow::modules::hash
+}  // namespace flow::module::hash