+++ plugins/modules/intern/island/island.hpp
@@ -5,7 +5,7 @@
 #include <logger2.h>
 #include <plugin2.h>
 
-namespace flow::modules::island {
+namespace flow::module::island {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger);
-}  // namespace flow::modules::island
+}  // namespace flow::module::island