+++ plugins/modules/intern/module.cpp
@@ -1,28 +1,30 @@
-#include "module.hpp"
+#include "../module.hpp"
 
-#include <hash.hpp>
-#include <island.hpp>
-#include <kerning.hpp>
-#include <regex.hpp>
-#include <toml.hpp>
-#include <trim.hpp>
-#include <vector.hpp>
+#include "hash/hash.hpp"
+#include "island/island.hpp"
+#include "text/kerning.hpp"
+#include "text/regex.hpp"
+#include "text/utf8.hpp"
+#include "toml/toml.hpp"
+#include "trim/trim.hpp"
+#include "vector/vector.hpp"
 
-namespace flow::modules {
+namespace flow::module {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger) {
     hash::init(host, logger);
-    vector::init(host, logger);
-    trim::init(host, logger);
     island::init(host, logger);
-    toml::init(host, logger);
-    text::regex::init(host, logger);
     text::kerning::init(host, logger);
+    text::regex::init(host, logger);
+    text::utf8::init(host, logger);
+    toml::init(host, logger);
+    trim::init(host, logger);
+    vector::init(host, logger);
 }
 
 void
 deinit() {
-    text::regex::deinit();
     text::kerning::deinit();
+    text::regex::deinit();
 }
-}  // namespace flow::modules
+}  // namespace flow::module