+++ plugins/modules/intern/text/intern/regex.cpp
@@ -1,4 +1,4 @@
-#include "regex.hpp"
+#include "../regex.hpp"
 
 #include <cstdint>
 #include <optional>
@@ -10,8 +10,8 @@
 
 #include <module2.h>
 
-#include <cache.hpp>
-#include <utilities.hpp>
+#include <intern/cache.hpp>
+#include <intern/string.hpp>
 
 namespace {
 namespace string = flow::string;
@@ -145,7 +145,7 @@ constinit SCRIPT_MODULE_TABLE info = {
 };
 }  // namespace
 
-namespace flow::modules::text::regex {
+namespace flow::module::text::regex {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *handle) {
     logger = handle;
@@ -158,4 +158,4 @@ void
 deinit() {
     patterns.reset();
 }
-}  // namespace flow::modules::text::regex
+}  // namespace flow::module::text::regex