+++ plugins/modules/intern/toml/toml.hpp
@@ -5,7 +5,7 @@
 #include <logger2.h>
 #include <plugin2.h>
 
-namespace flow::modules::toml {
+namespace flow::module::toml {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger);
-}  // namespace flow::modules::toml
+}  // namespace flow::module::toml