+++ plugins/modules/intern/trim/trim.hpp
@@ -5,7 +5,7 @@
 #include <logger2.h>
 #include <plugin2.h>
 
-namespace flow::modules::trim {
+namespace flow::module::trim {
 void
 init(HOST_APP_TABLE *host, LOG_HANDLE *logger);
-}  // namespace flow::modules::trim
+}  // namespace flow::module::trim