+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v0.2.0
+
+- 一部プロパティメニュー名の変更
+- プロパティメニュー`値を揃える`に`全ての区間`を追加
+- プロパティメニュー`値を反転`に`現在の区間`，`以前の区間`，`以降の区間`を追加
+- プロパティメニューに`エフェクトをコピー`を追加
+- `Motion@FlowType_K` の `Based On` が `Whole` 以外において，個別オブジェクト数が文字数の定数倍になる場合の挙動を追加
+- `Motion@FlowType_K` の `Based On` に `Objects` を追加 (個別オブジェクト毎を `Characters` からこれに変更)
+- `Kerning@FlowType_K` に影響度合いを指定する `Influence` を追加
+- `Kerning@FlowType_K` で個別オブジェクト数が文字数の定数倍になる場合の挙動を追加
+- `Motion@FlowType_K` の `Lyric` で退場時の伸縮が想定と逆になっていた問題を修正
+- トランスフォームのターゲットがローカル空間でもワールド空間でもない場合は変換処理が行われないように修正
+- `Align@FlowType_K` のフィルタプラグイン化
+- `Deform@FlowType_K` のフィルタプラグイン化
+- `Transform@FlowType_K` のフィルタプラグイン化
+- プリセットの追加
+
 ## v0.1.0
 
 - Release